+++ openeo/rest/datacube.py
@@ -981,10 +981,13 @@ class DataCube(_ProcessGraphAbstraction):
         :param boundary: Behavior to apply if the number of values for the axes x and y is not a
             multiple of the corresponding value in the size parameter.
             Options are:
-                pad (default): pad the data cube with the no-data value null to fit the required window size.
-                trim: trim the data cube to fit the required window size.
-            Set the parameter align to specifies to which corner the data is aligned to.
-        :param align: If the data requires padding or trimming (see parameter boundary), specifies
+
+                - ``pad`` (default): pad the data cube with the no-data value null to fit the required window size.
+                - ``trim``: trim the data cube to fit the required window size.
+
+            Use the parameter ``align`` to align the data to the desired corner.
+
+        :param align: If the data requires padding or trimming (see parameter ``boundary``), specifies
             to which corner of the spatial extent the data is aligned to. For example, if the data is
             aligned to the upper left, the process pads/trims at the lower-right.
         :param context: Additional data to be passed to the process.
@@ -1043,12 +1046,6 @@ class DataCube(_ProcessGraphAbstraction):
         the dimension labels will be incrementing integers starting from zero, which can be changed using
         rename_labels afterwards. The number of labels will equal to the number of values computed by the process.
 
-        .. note::
-            .. versionchanged:: 0.13.0
-                arguments ``code``, ``runtime`` and ``version`` are deprecated if favor of the standard approach
-                of using an :py:class:`UDF <openeo.rest._datacube.UDF>` object in the ``process`` argument.
-                See :ref:`old_udf_api` for more background about the changes.
-
         :param code: [**deprecated**] UDF code or process identifier (optional)
         :param runtime: [**deprecated**] UDF runtime to use (optional)
         :param process: the "child callback":
@@ -1075,6 +1072,12 @@ class DataCube(_ProcessGraphAbstraction):
 
         :return: A datacube with the UDF applied to the given dimension.
         :raises: DimensionNotAvailable
+
+        .. versionchanged:: 0.13.0
+            arguments ``code``, ``runtime`` and ``version`` are deprecated if favor of the standard approach
+            of using an :py:class:`UDF <openeo.rest._datacube.UDF>` object in the ``process`` argument.
+            See :ref:`old_udf_api` for more background about the changes.
+
         """
         # TODO #137 #181 #312 remove support for code/runtime/version
         if runtime or (isinstance(code, str) and "\n" in code):